proctor/few_shot/knn_implementation.py
"""
Advanced KNN implementation with proper embeddings and similarity calculation.

NOTE: This is a placeholder for future implementation. To use this module,
you'll need to install additional dependencies:

    uv pip install sentence-transformers scikit-learn

This module will be integrated with the KNN technique in a future update.
"""

from typing import List, Dict, Optional, Tuple
import numpy as np
from ..utils import log

# These imports would be needed for the actual implementation
try:
    from sentence_transformers import SentenceTransformer
    from sklearn.metrics.pairwise import cosine_similarity

    _DEPS_AVAILABLE = True
except ImportError:
    _DEPS_AVAILABLE = False
    log.warning(
        "Optional dependencies for KNN are not available. "
        "Install with: uv pip install sentence-transformers scikit-learn"
    )


class EmbeddingCache:
    """
    Cache for storing and retrieving text embeddings.
    """

    def __init__(self, max_size: int = 1000):
        """
        Initialize the embedding cache.

        Args:
            max_size (int): Maximum number of embeddings to store
        """
        self.cache: Dict[str, np.ndarray] = {}
        self.max_size = max_size

    def get(self, text: str) -> Optional[np.ndarray]:
        """
        Get embedding for a text if it exists in cache.

        Args:
            text (str): Text to retrieve embedding for

        Returns:
            Optional[np.ndarray]: Embedding if available, None otherwise
        """
        return self.cache.get(text)

    def add(self, text: str, embedding: np.ndarray) -> None:
        """
        Add embedding to cache.

        Args:
            text (str): Text
            embedding (np.ndarray): Embedding to cache
        """
        # Simple LRU implementation - remove oldest item if cache is full
        if len(self.cache) >= self.max_size and text not in self.cache:
            # Remove first item (oldest)
            self.cache.pop(next(iter(self.cache)))

        self.cache[text] = embedding


class SemanticKNN:
    """
    KNN implementation using semantic similarity with text embeddings.
    """

    def __init__(self, model_name: str = "all-MiniLM-L6-v2", cache_size: int = 1000):
        """
        Initialize the Semantic KNN.

        Args:
            model_name (str): Name of the sentence-transformer model to use
            cache_size (int): Maximum size of embedding cache

        Raises:
            ImportError: If required dependencies are not installed
        """
        if not _DEPS_AVAILABLE:
            raise ImportError(
                "Required dependencies not available. "
                "Install with: uv pip install sentence-transformers scikit-learn"
            )

        self.model = SentenceTransformer(model_name)
        self.cache = EmbeddingCache(max_size=cache_size)

    def _get_embedding(self, text: str) -> np.ndarray:
        """
        Get embedding for text, using cache if available.

        Args:
            text (str): Text to embed

        Returns:
            np.ndarray: Embedding vector
        """
        cached = self.cache.get(text)
        if cached is not None:
            return cached

        embedding = self.model.encode(text, convert_to_numpy=True)
        self.cache.add(text, embedding)
        return embedding

    def find_nearest(
        self,
        query: str,
        candidates: List[Dict[str, str]],
        text_key: str = "input",
        k: int = 3,
    ) -> List[Tuple[Dict[str, str], float]]:
        """
        Find k-nearest neighbors to a query from a list of candidates.

        Args:
            query (str): Query text
            candidates (List[Dict[str, str]]): List of candidate examples
            text_key (str): Key in candidate dict to use for comparison
            k (int): Number of neighbors to return

        Returns:
            List[Tuple[Dict[str, str], float]]: List of (candidate, similarity) pairs
        """
        if not candidates:
            return []

        # Get query embedding
        query_embedding = self._get_embedding(query)

        # Get embeddings for all candidates
        candidate_texts = [c[text_key] for c in candidates]
        candidate_embeddings = np.vstack(
            [self._get_embedding(text) for text in candidate_texts]
        )

        # Calculate similarities
        similarities = cosine_similarity(
            query_embedding.reshape(1, -1), candidate_embeddings
        )[0]

        # Sort by similarity and take top k
        top_indices = similarities.argsort()[-k:][::-1]

        # Return candidates with their similarity scores
        return [(candidates[i], similarities[i]) for i in top_indices]


# Example usage:
"""
# First, install the required dependencies
# uv pip install sentence-transformers scikit-learn

from proctor.few_shot.knn_implementation import SemanticKNN

# Initialize the KNN
knn = SemanticKNN()

# Create a pool of examples
examples_pool = [
    {"input": "What is machine learning?", "output": "A field of AI..."},
    {"input": "Explain quantum computing", "output": "Quantum computing..."},
    # ... more examples
]

# Find the most similar examples to a query
query = "How does deep learning work?"
similar_examples = knn.find_nearest(query, examples_pool)

# Use these examples in your prompt
for example, similarity in similar_examples:
    print(f"Similarity: {similarity:.2f}")
    print(f"Example: {example['input']}")
    print(f"Answer: {example['output']}\n")
"""
