pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "proctor-ai"
version = "0.1.1"
authors = [
  {name = "Svngoku", email = "svngokuofficiel@gmail.com"},
]
description = "A comprehensive Python package for structured prompt engineering techniques"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["prompt-engineering", "llm", "ai", "nlp", "gpt", "openai", "prompt-templates"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
]
dependencies = [
    "build>=1.2.2.post1",
    "litellm>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=0.20.0",
    "rich>=13.0.0",
    "twine>=6.1.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
knn = [
    "sentence-transformers>=2.0.0",
    "scikit-learn>=1.0.0"
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.270",
    "build>=0.10.0"
]
all = [
    "sentence-transformers>=2.0.0",
    "scikit-learn>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.270",
    "build>=0.10.0"
]

[project.urls]
"Homepage" = "https://github.com/svngoku/proctor"
"Bug Tracker" = "https://github.com/svngoku/proctor/issues"
"Documentation" = "https://github.com/svngoku/proctor/blob/master/README.md"
"Source Code" = "https://github.com/svngoku/proctor" 
